Integrated React using hacky vanillajs & timeout
