+++ index.html
@@ -19,6 +19,10 @@
     </noscript>
     <div id="app"></div>
     <script type="module" src="/src/docs/main.ts"></script>
+
+    <!-- Adding React for Framework Preview -->
+    <script type="module" src="/src/react/src/frameworkpage.tsx"></script>
+
     <!-- Google tag (gtag.js) -->
     <script async src="https://www.googletagmanager.com/gtag/js?id=G-14PMXQVXTC"></script>
     <script>
