+++ src/react/index.html
@@ -6,7 +6,7 @@
     <title>Slim Select React</title>
   </head>
   <body>
-    <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
+    <div id="react-root"></div>
+    <script type="module" src="/src/frameworkpage.tsx"></script>
   </body>
 </html>
