+++ src/react/package.json
@@ -3,6 +3,16 @@
   "private": false,
   "description": "React integration",
   "version": "1.0.0",
+  "keywords": [
+    "react"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/brianvoe/slim-select.git"
+  },
+  "author": "Brian Voelker <brian@webiswhatido.com> (http://webiswhatido.com)",
+  "homepage": "https://slimselectjs.com/frameworks#vue",
+  "license": "MIT",
   "type": "module",
   "scripts": {
     "dev": "vite",
@@ -51,5 +61,8 @@
   "main": "dist/slimselectreact.ssr.js",
   "browser": "dist/slimselect.global.js",
   "module": "dist/slimselectreact.es.js",
-  "types": "dist/slimselect.react.d.ts"
+  "types": "dist/slimselect.react.d.ts",
+  "files": [
+    "dist"
+  ]
 }
